game-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus-game-engine"
version = "0.1.0"
description = "Project Nexus - AI-driven geopolitical simulation game engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic",
    "python-dotenv",
    "redis",
    "asyncpg",
    "sqlalchemy[asyncio]",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]
